+++ Backend/error.log
@@ -1,0 +1,10 @@
+Route not found /pepito 2023-03-24T13:32:50.418Z
+Route not found /pepito 2023-03-24T13:33:22.936Z
+Route not found /pepito 2023-03-24T13:42:57.353Z
+Route not found /eventos 2023-03-24T14:01:18.026Z
+Route not found /eventos 2023-03-24T14:03:40.628Z
+Route not found /eventos 2023-03-24T14:03:48.803Z
+Route not found /eventos 2023-03-24T14:03:56.145Z
+Route not found /events/city/Madrid 2023-03-24T14:03:59.053Z
+Route not found /eventos 2023-03-24T14:04:47.525Z
+Route not found /evemts/eventos 2023-03-24T14:05:03.830Z
