+++ Backend/models/Event.js
@@ -4,7 +4,7 @@ const Schema = mongoose.Schema;
 
 const eventSchema = new Schema(
     {
-        id: { type: Number },
+        id: { type: String },
         title: { type: String, required: true },
         category: {type: String },
         location: {type: String },
