+++ Backend/routes/event.routes.js
@@ -1,0 +1,159 @@
+const express = require('express');
+const mongoose = require('mongoose');
+
+const Event = require('../models/Event');
+
+const router = express.Router();
+
+
+// QUERYS
+
+    //Events List
+router.get('/', async (req, res, next) => {
+
+    try {
+    const { limit, offset } = req.query;
+
+    return Event.find()
+    .then(events => {
+        if ( offset && limit ){
+        return res.status(200).json(events.slice(offset, parseInt(limit) + parseInt(offset)));
+        } else if (offset) {
+        return res.status(200).json(events.slice(offset));
+        } else if (limit) {
+        return res.status(200).json(events.slice(0, limit));
+        }
+        return res.status(200).json(events);
+    });
+    } catch (error) {
+        return next(error);
+    }
+});
+
+    //Events by date
+router.get('/date', async (req, res, next) => {
+    const { dtstart} = req.query;
+
+    try { 
+        const eventbyDate = await Event.find({$and: [{dtstart : { $lte: dtstart}} , {dtend: {$gte: dtstart}} ]} );
+        if ( eventbyDate.length ){
+        return res.status(200).json(eventbyDate);
+        } else {
+            return res.status(404).json('No event found by this date')
+        }
+    } catch (err) {
+        return next(error);
+    }
+});
+
+    //Events from DT
+router.get('/dates', async (req, res, next) => {
+    const { dtstart, dtend } = req.query;
+
+    try {
+        let events = [];
+        if ( dtstart && dtend ) {
+            events = await Event.find( {dtstart : { $lte: dtstart}, dtend: {$gte: dtend}  });
+        }else {
+            events = await Event.find();
+        }
+        return res.status(200).json(events);
+
+    } catch (err) {
+        return next(error);
+    }
+});
+
+// PARAMS
+
+    //Events by ID
+router.get('/:id', async (req, res, next) => {
+    const id = req.params.id;
+    try { 
+        const event = await Event.findById(id);
+        if ( event ) {
+        return res.status(200).json(event);
+        } else {
+            return res.status(404).json('No event found by this id')
+        }
+    } catch (err) {
+        return next(error);
+    }
+});
+
+    //Events by CITY
+router.get('/city/:city', async (req, res, next) => {
+    const { city } = req.params;
+
+    try { 
+        const eventbyCity = await Event.find( {city: city});
+        if (eventbyCity ){
+        return res.status(200).json(eventbyCity);
+        } else {
+            return res.status(404).json('No event found in this city')
+        }
+    } catch (err) {
+        return next(error);
+    }
+});
+
+// union query params
+//router.get('/events/:idOrCity', async (req, res) => {
+//    const { idOrCity  } = req.params;
+//
+//    try { 
+//        let event = null;
+//        try {
+//            const eventId = mongoose.Types.ObjectId(idOrCity);
+//            event = await Event.findById(eventId);
+//        } catch (err) {
+//            if (err.name === 'BSONTypeError') {
+//                event = await Event.findOne( {city: idOrCity} );
+//            } else {
+//                return res.status(500).json(err);
+//            }
+//        }
+//
+//        if (event) {
+//            return res.status(200).json(event);
+//        } else {
+//            return res.status(404).json('No Event found by this Id');
+//        }
+//    } catch (err) {
+//        return res.status(500).json(err);
+//    }
+//});  
+
+
+//POST 
+
+// POST
+router.post('/', async (req, res, next) => { 
+    const { title, category, location , city, province, lat, long, dtstart, dtend, price, info,  link, artist, img } = req.body;
+    const event = {
+        title, 
+        category, 
+        location , 
+        city, 
+        province, 
+        lat, 
+        long, 
+        dtstart, 
+        dtend, 
+        price, 
+        info,  
+        link, 
+        artist, 
+        img
+    };
+    try{
+        const newEvent = new Event(event);
+        const eventCreated = await newEvent.save();        
+        return res.status(201).json(eventCreated);
+    } catch (error) {
+        next(error);
+    }
+});
+
+
+module.exports = router;
