+++ Backend/utils/db.js
@@ -2,11 +2,13 @@ const mongoose = require('mongoose');
 
 const DB_URL = 'mongodb://localhost:27017/Proyecto-Final';
 
-mongoose.connect(DB_URL, {
+const connect = () => mongoose.connect(DB_URL, {
     useNewUrlParser: true,
     useUnifiedTopology: true
 });
 
 module.exports = {
     DB_URL,
+    connect
   };
+
